Set up boiler plater for image agent
Set up the ImageModel and made it be consumed by image_agent/player.py,
so that we can play the tournament.
Set up a Controller abstract class that allows have various
implementations.
Rename det_647_e114.th to det.th so that we don't have crashes if we
forget to rename it.